Abstract away mintTransferableNft
